network improvements, change most messages to Uint8

+++ server/src/room/browser/message/BrowserActiveMessage.js
@@ -5,13 +5,13 @@ export default class DefaultActiveMessage {
         let _room = sender.server.getRoom(data.id);
         if (_room) {
             if (!_room.password || data.password === _room.password) {
-                sender.send(new Float64Array([MSG.JOIN_SUCCEEDED]));
+                sender.send(new Uint8Array([MSG.JOIN_SUCCEEDED]));
                 sender.room = _room;
             } else {
-                sender.send(new Float64Array([MSG.JOIN_FAILED, MSG.INCORRECT_PASSWORD]));
+                sender.send(new Uint8Array([MSG.JOIN_FAILED, MSG.INCORRECT_PASSWORD]));
             }
         } else {
-            sender.send(new Float64Array([MSG.JOIN_FAILED, MSG.ROOM_DOES_NOT_EXIST]));
+            sender.send(new Uint8Array([MSG.JOIN_FAILED, MSG.ROOM_DOES_NOT_EXIST]));
         }
     }
 
@@ -21,9 +21,9 @@ export default class DefaultActiveMessage {
 
     static 'name'(room, sender, data) {
         if (data.length > 8) {
-            sender.send(new Float64Array([MSG.NAME_INVALID]));
+            sender.send(new Uint8Array([MSG.NAME_INVALID]));
         } else {
-            sender.send(new Float64Array([MSG.NAME_SET_SUCCESSFULLY]));
+            sender.send(new Uint8Array([MSG.NAME_SET_SUCCESSFULLY]));
             sender.name = data;
         }
     }
